Upgrade lusca package (w/ CSRF header fix) from 1.0.x to 1.3.x. Remove crypton checks on Lusca object.
Prior to this change it was not possible to use the full gammut of config
options for the Lusca security headers package.  For example you could not
pass 'false' to one of the types to disable that type (e.g CSP or HSTS).

Some types just were not validated at all.

Control for where and how to set security headers should always be the
responsibility of the developer. This is especially true for settings like
HSTS which can have severe unintended consequences if configured incorrectly.

In the sample config I added the other Object keys that Lusca accepts
and provided what should be sane defaults to make it easier for a dev to see
what is possible.

Also emitting the current securityHeaders config to STDOUT when the server is run.

I ran the concept of this change by @daviddahl before committing via Twitter DM.