jshint cleanup. Add a default global .jshint config. Fix low hanging fruit.
I use the jshint plugin in Atom editor. .jshint will work with any jshint
tool though to provide a consistent set of standards for the project.

In this cleanup primarily == to === and !== to !===.

There are still a number of jshint errors which need to be looked at in
more detail.

See:
http://javascript.crockford.com/style2.html
https://javascriptweblog.wordpress.com/2011/02/07/truth-equality-and-javascript/

All tests (client and server) pass.